+++ CHANGELOG.md
@@ -21,6 +21,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 * [Add path to ASTCollecting](https://github.com/ionide/FSharp.Analyzers.SDK/pull/171) (thanks @nojaf!)
 * [Use Microsoft.Extensions.Logging instead of printf based logging infrastructure](https://github.com/ionide/FSharp.Analyzers.SDK/pull/175) (thanks @dawedawe!)
+* [Console finer colorized pretty printing and link to analyzer URI](https://github.com/ionide/FSharp.Analyzers.SDK/pull/180) (thanks @smoothdeveloper!)
 
 ## [0.21.0] - 2023-11-22